requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
python-multipart
httpx